requirements.txt
aiohttp==3.7.4.post0
argon2-cffi==20.1.0
arrow==1.1.0
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bleach==3.3.0
blinker==1.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.0
cloudpickle==1.6.0
colorama==0.4.4
commonmark==0.9.1
decorator==5.0.9
defusedxml==0.7.1
docutils==0.17.1
doit==0.33.1
entrypoints==0.3
ghp-import==1.1.0
hsluv==5.0.2
html5lib==1.1
idna==2.10
ipykernel==5.5.5
ipython==7.23.1
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.0
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
lxml==4.6.3
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==2.0.0
matplotlib-inline==0.1.2
micawber==0.5.3
mistune==0.8.4
multidict==5.1.0
natsort==7.1.1
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
Nikola==8.1.3
notebook==6.3.0
packaging==20.9
pandocfilters==1.4.3
parso==0.8.2
pep517==0.10.0
pexpect==4.8.0
phpserialize==1.3
pickleshare==0.7.5
piexif==1.1.3
Pillow==8.2.0
pip-tools==6.1.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pycparser==2.20
pygal==2.4.0
Pygments==2.9.0
pyinotify==0.9.6
pyparsing==2.4.7
Pyphen==0.10.0
pyrsistent==0.17.3
PyRSS2Gen==1.1
python-dateutil==2.8.1
python-frontmatter==1.0.0
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
rich==10.2.0
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
Send2Trash==1.5.0
six==1.16.0
smartypants==2.0.1
terminado==0.9.5
testpath==0.4.4
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typing-extensions==3.10.0.0
typogrify==2.0.7
Unidecode==1.2.0
urllib3==1.26.4
watchdog==2.1.1
wcwidth==0.2.5
webencodings==0.5.1
Yapsy==1.12.2
yarl==1.6.3
